# server/docs/swagger.yaml
definitions:
...
  model.Access:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      route:
        type: string
      updated_at:
        type: string
    type: object
...
  model.AlipayPreCreatePayToFrontend:
    properties:
      qr_code:
        type: string
    type: object
...
  model.CasbinItem:
    properties:
      method:
        description: 方法
        type: string
      path:
        description: 路径
        type: string
    type: object
  model.Coupon:
    properties:
      created_at:
        type: string
      discount_rate:
        type: number
      expired_at:
        type: string
      goods:
        items:
          $ref: '#/definitions/model.Goods'
        type: array
      id:
        type: integer
      limit:
        type: integer
      min_amount:
        type: number
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  model.EpayPreCreatePay:
    properties:
      money:
        description: "商品金额,如：1.00\t单位：元，最大2位小数"
        type: string
      name:
        description: 商品名称,如超过127个字节会自动截取
        type: string
      notify_url:
        description: 服务器异步通知地址
        type: string
      out_trade_no:
        description: 商户订单号
        type: string
      pid:
        description: 商户ID
        type: integer
      return_url:
        description: 页面跳转通知地址
        type: string
      sign:
        description: 签名字符串，所有参数按照参数名ASCII码从小到大排序（a-z），sign、sign_type、和空值不参与签名！sign
          = md5 ( a=b&c=d&e=f + KEY ) （注意：+ 为各语言的拼接符，不是字符！），md5结果为小写。
        type: string
      sign_type:
        description: 签名类型
        type: string
      type:
        description: 支付方式
        type: string
    type: object
  model.EpayPreCreatePayToFrontend:
    properties:
      epay_api_url:
        type: string
      epay_pre_create_pay:
        $ref: '#/definitions/model.EpayPreCreatePay'
    type: object
...
  model.FieldParamsItem:
    properties:
      condition:
        description: = > < <> like
        type: string
      condition_value:
        type: string
      field:
        type: string
      field_type:
        type: string
      operator:
        description: AND OR
        type: string
    type: object
...
  model.Goods:
    properties:
      coupon:
        items:
          $ref: '#/definitions/model.Coupon'
        type: array
      cover_image:
        type: string
      created_at:
        type: string
      deliver_text:
        type: string
      deliver_type:
        type: string
      des:
        type: string
      enable_traffic_reset:
        type: boolean
      goods_order:
        description: 基础参数
        type: integer
      goods_type:
        type: string
      id:
        type: integer
      is_renew:
        type: boolean
      is_sale:
        type: boolean
      is_show:
        type: boolean
      node_connector:
        type: integer
      node_speed_limit:
        type: integer
      nodes:
        description: 关联
        items:
          $ref: '#/definitions/model.Node'
        type: array
      price:
        description: todo 前端修改
        type: string
      price_3_month:
        description: 订阅参数
        type: string
      price_6_month:
        type: string
      price_12_month:
        type: string
      price_unlimited_duration:
        type: string
      quota:
        type: integer
      recharge_amount:
        description: 充值参数
        type: string
      stock:
        type: integer
      subject:
        type: string
      total_bandwidth:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      component:
        type: string
      created_at:
        type: string
      id:
        type: integer
      meta:
        $ref: '#/definitions/model.Meta'
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      remarks:
        type: string
      roles:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      updated_at:
        type: string
    type: object
  model.Meta:
    properties:
      icon:
        type: string
      isAffix:
        type: boolean
      isHide:
        type: boolean
      isIframe:
        type: boolean
      isKeepAlive:
        type: boolean
      isLink:
        type: string
      title:
        type: string
    type: object
...
  model.Node:
    properties:
      access:
        description: 访问控制
        items:
          $ref: '#/definitions/model.Access'
        type: array
      address:
        type: string
      aid:
        type: integer
      allowInsecure:
        type: boolean
      alpn:
        type: string
      created_at:
        type: string
      dest:
        type: string
      enabled:
        type: boolean
      encryption:
        type: string
      flow:
        type: string
      fp:
        type: string
      goods:
        description: 关联参数
        items:
          $ref: '#/definitions/model.Goods'
        type: array
      host:
        type: string
      hy_down_mbps:
        type: integer
      hy_obfs:
        type: string
      hy_obfs_password:
        type: string
      hy_ports:
        type: string
      hy_up_mbps:
        type: integer
      id:
        type: integer
      mode:
        type: string
      network:
        type: string
      node_order:
        type: integer
      node_speed_limit:
        type: integer
      node_type:
        type: string
      path:
        type: string
      pbk:
        type: string
      port:
        type: integer
      private_key:
        type: string
      protocol:
        description: 一些协议参数
        type: string
      remarks:
        description: 基础参数
        type: string
      scy:
        type: string
      security:
        type: string
      server_key:
        type: string
      service_name:
        type: string
      sid:
        type: string
      sni:
        type: string
      spx:
        type: string
      total_down:
        description: Byte
        type: integer
      total_up:
        description: 已用上行/已用下行，统计节点流量时使用
        type: integer
      traffic_rate:
        type: number
      transfer_address:
        description: 中转参数
        type: string
      transfer_node_id:
        type: integer
      transfer_port:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      uuid:
        description: 共享节点需要的uuid;订阅下发是实际的用户uuid
        type: string
      v:
        type: string
    type: object
  model.NodeSharedReq:
    properties:
      url:
        type: string
    type: object
...
  model.Order:
    properties:
      balance_amount:
        type: string
      buyer_logon_id:
        type: string
      buyer_pay_amount:
        type: string
      coupon_amount:
        type: string
      coupon_id:
        type: integer
      coupon_name:
        type: string
      created_at:
        type: string
      customer_service_id:
        description: 服务参数
        type: integer
      deliver_text:
        type: string
      deliver_type:
        type: string
      des:
        type: string
      duration:
        type: integer
      goods_id:
        description: 商品参数
        type: integer
      goods_type:
        type: string
      id:
        type: integer
      order_remarks:
        type: string
      order_type:
        type: string
      original_amount:
        type: string
      out_trade_no:
        type: string
      pay_id:
        description: 支付参数
        type: integer
      pay_info:
        allOf:
        - $ref: '#/definitions/model.PreCreatePayToFrontend'
        description: 支付信息，epay，alipay等"
      pay_type:
        type: string
      price:
        type: string
      subject:
        type: string
      total_amount:
        type: string
      trade_no:
        type: string
      trade_status:
        type: string
      updated_at:
        type: string
      user_id:
        description: 关联用户
        type: integer
      user_name:
        type: string
    type: object
  model.Pagination:
    properties:
      order_by:
        type: string
      page_num:
        type: integer
      page_size:
        type: integer
    type: object
...
  model.PreCreatePayToFrontend:
    properties:
      alipay_info:
        $ref: '#/definitions/model.AlipayPreCreatePayToFrontend'
      epay_info:
        $ref: '#/definitions/model.EpayPreCreatePayToFrontend'
    type: object
...
  model.QueryParams:
    properties:
      field_params_list:
        items:
          $ref: '#/definitions/model.FieldParamsItem'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/model.Pagination'
        description: 分页参数
      table_name:
        type: string
    required:
    - table_name
    type: object
...
  model.Role:
    properties:
      casbins:
        items:
          $ref: '#/definitions/model.CasbinItem'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menus:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      role_name:
        type: string
      status:
        type: boolean
      updated_at:
        type: string
      user_group:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      balance:
        type: number
      created_at:
        type: string
      enable:
        type: boolean
      enable_email:
        type: boolean
      enable_tg_bot:
        description: 通知参数
        type: boolean
      enable_web_mail:
        type: boolean
      id:
        type: integer
      invitation_code:
        type: string
      nick_name:
        type: string
      orders:
        description: 订单，has many
        items:
          $ref: '#/definitions/model.Order'
        type: array
      password:
        type: string
      referrer_user_id:
        type: integer
      role_group:
        description: 关联参数
        items:
          $ref: '#/definitions/model.Role'
        type: array
      tg_id:
        type: integer
      updated_at:
        type: string
      user_name:
        type: string
      when_balance_changed:
        type: boolean
      when_purchased:
        type: boolean
      when_service_almost_expired:
        type: boolean
    type: object
...
  response.ResponseStruct:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact:
    url: https://github.com/shuangyu233/AirGo_Modify
  description: AirGo前后分离,多用户代理面板
  license:
    name: GPL v3.0
    url: https://github.com/shuangyu233/AirGo_Modify/blob/main/LICENSE
  title: AirGo
paths:
...
  /api/admin/menu/delMenu:
    delete:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 删除菜单
      tags:
      - '[admin api] menu'
  /api/admin/menu/getAllMenuList:
    get:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取全部菜单列表
      tags:
      - '[admin api] menu'
  /api/admin/menu/newMenu:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 新建菜单
      tags:
      - '[admin api] menu'
  /api/admin/menu/updateMenu:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 修改菜单
      tags:
      - '[admin api] menu'
...
  /api/admin/node/createx25519:
    get:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: reality x25519
      tags:
      - '[admin api] node'
  /api/admin/node/deleteNode:
    delete:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Node'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 删除节点
      tags:
      - '[admin api] node'
  /api/admin/node/getNodeList:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取节点列表
      tags:
      - '[admin api] node'
  /api/admin/node/getNodeListWithTraffic:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取节点列表，带流量信息
      tags:
      - '[admin api] node'
  /api/admin/node/getNodeServerStatus:
    get:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取节点服务器状态
      tags:
      - '[admin api] node'
  /api/admin/node/newNode:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Node'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 新建节点
      tags:
      - '[admin api] node'
  /api/admin/node/newNodeShared:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/model.Node'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 新增共享节点
      tags:
      - '[admin api] node'
  /api/admin/node/nodeSort:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/model.Node'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 节点排序
      tags:
      - '[admin api] node'
  /api/admin/node/parseUrl:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.NodeSharedReq'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 解析
      tags:
      - '[admin api] node'
  /api/admin/node/updateNode:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Node'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 更新节点
      tags:
      - '[admin api] node'
  /api/admin/order/getOrderList:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取全部订单，分页获取
      tags:
      - '[admin api] order'
  /api/admin/order/orderSummary:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取订单统计
      tags:
      - '[admin api] order'
  /api/admin/order/updateOrder:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Order'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 更新用户订单
      tags:
      - '[admin api] order'
...
swagger: "2.0"